Sort out the calc config
The default value for calc-settings-file is "~/.emacs.d/calc.el" but
unfortunately I have "~/.emacs.d/" on my load path, so it was shadowing
the *real* calc.el, with the result that calc didn't work at all.  So I
changed it to "calc-settings.el" and now all is well again.