Fix the color identifier setup
Now properly remove the existing colors from the syntax highlighting and
replace it with judicious use of bold and italics.  Note that I use
`face-map-add-relative` for this, but I could never work out the right
syntax for setting multiple specs at once, so I do it in bits.

This is then added to the hook for all programming language modes and
org mode.