Adaptive soft wrap
Found some code on emacs Wiki to do this, but it is based on old
filladapt.el, so not sure how stable it is

I got it for swift files, and it worked for one but not the other - not
sure why.  Requires you to hit return on a line before it kicks in.