Make org-date face inherit from org-table
This is because org-table is a compressed font, but the font lock
keywords have org-date overriding org-table face for dates that are
inside tables, which messes up the column alignment.

It might be possible to fix this by modifying org-set-font-lock-defaults
(for instance using prepend instead of t for the OVERRIDE argument) but
that looks tricky.  Instead, I just make the dates narrow everywhere.
Outside of tables, it shouldn't matter