Use helpful.el for better `*help*` buffers
Also customize the `button` face in a few ways: remove the underline
because it is just too close to the baseline (can't see any way to fix
that), which cuts off descenders in g, j, y, etc.  Make it semi-bold
and slanting, but change the color to light gray to offset the
increased boldness.