Sort out the company auto complete configuration
This was mainly a problem in python files, where I was frequently
getting extraneous newlines inserted for no reason in the middle of
words. I do not know why that was happening, but I have worked around
the problem by changing the frontend to use the echo area instead of a
pop-up list.

While investigating this, I came across
corfu.el (https://github.com/minad/corfu), which may be a better
solution to try someday.

I also disabled the auto-complete.el package since it is obsolete,
although it turned out not to have been causing any real trouble.