Try out detached.el for long-running background processes
This is sort of working, but it needs more testing. I am unconvinced
that it is the best approach. I remember a previous package I used to
use, which seemed more intuitive.

Anyway, I also installed Jogn Wiegley's alert.el since the sbus
integration of detached.el was not working for me. I have it set up to
send alerts via applescript to the macos notification center.